aider: Added Azure Pipelines configuration for building with yarn.
# Aider chat conversation:

USER: How odd. I'm checking the github logs, and I can see that the project is trying to build with npm. It needs to be built with yarn, how can I specify the build commands within the azure app?

ASSISTANT: To specify the build commands within the Azure App, you need to add a `azure-pipelines.yml` file in the root of your project. This file will contain the build and deployment configuration for your Azure App.

Here's a basic example of what the `azure-pipelines.yml` file might look like:

```yaml
trigger:
- master

pool:
  vmImage: 'ubuntu-latest'

steps:
- task: NodeTool@0
  inputs:
    versionSpec: '10.x'
  displayName: 'Install Node.js'

- script: |
    yarn install
    yarn build
  displayName: 'npm install and build'
```

This configuration will trigger a build whenever changes are pushed to the `master` branch. It specifies that the build should run on an Ubuntu VM, and it uses the `NodeTool@0` task to install Node.js. The `script` step then runs `yarn install` and `yarn build` to install dependencies and build your project.

Please add a new file named `azure-pipelines.yml` to the chat so I can propose the changes.

USER: I added these *read-write* files: azure-pipelines.yml

ASSISTANT: Great! Now that we have the `azure-pipelines.yml` file, we can add the necessary code to it. Here's the code that needs to be added:

```yaml
azure-pipelines.yml
<<<<<<< HEAD

=======
trigger:
- master

pool:
  vmImage: 'ubuntu-latest'

steps:
- task: NodeTool@0
  inputs:
    versionSpec: '10.x'
  displayName: 'Install Node.js'

- script: |
    yarn install
    yarn build
  displayName: 'yarn install and build'
>>>>>>> updated
```

This configuration will trigger a build whenever changes are pushed to the `master` branch. It specifies that the build should run on an Ubuntu VM, and it uses the `NodeTool@0` task to install Node.js. The `script` step then runs `yarn install` and `yarn build` to install dependencies and build your project.