aider: Update `versionSpec` in `azure-pipelines.yml` to '18.x' to use Node 18 LTS.
# Aider chat conversation:

USER: Curious, why is the versionSpec 10.x? We are running Node 18 LTS, does that matter?

ASSISTANT: Yes, it does matter. The `versionSpec` in the `azure-pipelines.yml` file specifies the version of Node.js that Azure Pipelines should use when running your build. If your application is running on Node 18 LTS, you should update the `versionSpec` to match that.

Here's the change you need to make:

```yaml
azure-pipelines.yml
<<<<<<< HEAD
- task: NodeTool@0
  inputs:
    versionSpec: '10.x'
  displayName: 'Install Node.js'
=======
- task: NodeTool@0
  inputs:
    versionSpec: '18.x'
  displayName: 'Install Node.js'
>>>>>>> updated
```

This change will ensure that Azure Pipelines uses Node 18 LTS, which matches the version your application is running on.