aider: Created SVG favicon with text "AI".
# Aider chat conversation:

USER: We don't seem to have a favicon.ico. Can let' just create an SVG of text "AI".

ASSISTANT: Sure, we can create a new SVG file with the text "AI" as the favicon. Here's the SVG code for that:

```svg
public/favicon.svg
<<<<<<< HEAD

=======
<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 50 50">
  <text x="5" y="45" font-size="40">AI</text>
</svg>
>>>>>>> updated
```

After creating the SVG file, you need to link it in your HTML files (usually in the `<head>` section) like this:

```html
<link rel="icon" type="image/svg+xml" href="/favicon.svg">
```

You would need to add this line to all your HTML files or to a common layout file if you have one. If you're using a framework like Next.js, you would add this line to the `_document.js` or `_document.tsx` file. If you need help with this, please provide the relevant file(s).