aider: Update favicon to use Inter font.
# Aider chat conversation:

USER: Let's update our favicon to use Inter font.

ASSISTANT: To use the Inter font in the favicon, we need to modify the SVG file that represents the favicon. We will add the font-family attribute to the text element in the SVG.

Here is the change:

```svg
public/favicon.svg
<<<<<<< HEAD
<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 50 50">
  <text x="5" y="45" font-size="40">AI</text>
</svg>
=======
<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 50 50">
  <text x="5" y="45" font-size="40" font-family="Inter">AI</text>
</svg>
>>>>>>> updated
```

Please note that this change will only apply if the Inter font is available on the client's machine or is loaded from a web font service. If the Inter font is not available, the client's browser will use a fallback font.